impl: Java Time API,  Dates changed with LocalDateTime, Long variables changed with Integer variables

+++ src/main/java/vip/yazilim/p2g/web/service/p2g/relation/ISongService.java
@@ -20,7 +20,7 @@ import java.util.Optional;
  */
 public interface ISongService extends ICrudService<Song, String> {
 
-    Song addSongToRoom(String roomUuid, String songId, String songUri, String songName, Long durationMs, int votes) throws DatabaseException, InvalidArgumentException;
+    Song addSongToRoom(String roomUuid, String songId, String songUri, String songName, Integer durationMs, int votes) throws DatabaseException, InvalidArgumentException;
 
     Optional<Song> getPausedSong(String roomUuid) throws DatabaseReadException;
